requirements.txt
fastapi
uvicorn
requests
apscheduler
sqlalchemy
pymysql
python-dotenv
python-dateutil
newspaper3k